♻️ Use 64 byte specific keccak instruction,
The `keccak256-aarch64-simd` crate includes a function that can accept variable-lengthed inputs, however this function has a bit of overhead due to manual padding. The input size to `keccak256` is fixed to 64 bytes here, so we can take advantage of the `simd_keccak_64b_single` function to reduce padding overhead.